Refuse all jobs
This is a breaking change, towards a read-only package backend.

Suggested next step is to refactor the packages and processors routers
to drop processing responsibilities and only provide metadata and data
access.

The main justifications for this change are to produce and publish
versioned data packages in a reliable way offline, in an extract/analysis
pipeline that can run in batch mode, and reduce online server load and
app/package complexity.